Module db_sqlx_sqlite::prelude

Re-exports

pub use super::*;

Modules

dev

useful imports for supporting a new database

errors

represents all the ways a trait can fail using this crate

ops

meta operations like migration and connecting to a database

tests

Test utilities

Structs

CreateGist

Data required to create a gist in DB creation date defaults to time at which creation method is called

CreateGistComment

Data required to create a comment on a Gist creation date defaults to time at which creation method is called

Creds

data structure describing credentials of a user

EmailRegisterPayload

payload to register a user with username and email

Gist

Represents a gist

GistComment

Represents a comment on a Gist

Password

data structure containing only a password field

UpdateEmailPayload

payload to update email in the database

UpdateUsernamePayload

payload to update a username in database

UsernameRegisterPayload

payload to register a user with only username

Enums

DBError

Error data structure grouping various error subtypes

GistVisibility

Gist visibility

Traits

CloneGPDatabse

Trait to clone GPDatabse

Connect

Create databse connection

DBOps

Database operations trait(migrations, pool creation and fetching connection from pool)

GPDatabse

foo

GetConnection

Get database connection

Migrate

database migrations

Type Definitions

BoxDynError

Convenience type alias for grouping driver-specific errors

DBResult

Generic result data structure